1) написал методы для поиска простого маршрута (еще не тестил)
2) изменил метод генерации

3) сделал json для postman, но он не импортируется

+++ src/main/java/com/example/prac/data/model/Ticket.java
@@ -56,5 +56,5 @@ public class Ticket {
     private LocalTime arrivalTime;
 
     @Column(name = "hours", nullable = false)
-    private int hours;
+    private double hours;
 }
